==> Graph Neural Network Modeling/03_download data.py <==
# -*- coding: utf-8 -*-
# @File       : 03_download data.py
# @Author     : Yuchen Chai
# @Date       : 2022/12/24 17:29
# @Description:

# Copy the user list to StravaData\Fetch_Network
# Rename the list file to users.csv

# StavaData\Fetch_Network\run_get user friend.py
