==> Graph Neural Network Modeling/04_copy aggregated data.py <==
# -*- coding: utf-8 -*-
# @File       : 04_copy aggregated data.py
# @Author     : Yuchen Chai
# @Date       : 2022/12/27 9:53
# @Description:

# Copy StravaData\Fetch_Network\aggregated.csv to 2_pipeline\130_strava network experiment\store\xxxx aggregated.csv
